backend/requirements.txt
django==4.2.7
psycopg2-binary==2.9.9
django-allauth==0.58.2
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.26.5
django-cors-headers==4.3.1